+++ src/main/java/myapp/SzakdolgozatBE/episode/EpisodeResource.java
@@ -72,7 +72,7 @@ public class EpisodeResource {
     }
     
     //todo
-    public void changeRating(long id, double rating) {
+    public void changeRating(long id, int rating) {
         service.changeRating(id, rating);
     }
 }
